Revert "Restructured Components files/folders and naming conventions"
