+++ client/src/App.js
@@ -1,6 +1,6 @@
 import React from "react";
 import ReactDOM, { render } from "react-dom";
-import BasePage from "./components/pages/basePage/BasePage";
+import BasePage from "./components/index/BasePage";
 //import '../public/index.html'
 
 import "./css/styles.css";
