--- client/src/components/common/footer/footer.js
@@ -1,11 +1,0 @@
-import React, { Component } from "react";
-
-export default class HeaderBlock extends Component {
-  render() {
-    return (
-      <di>
-        <div className="row" className="footer">Footer</div>
-      </di>
-    );
-  }
-}
