+++ client/src/components/common/logo.js
